+++ mods/src/multiplayer/internal.h
@@ -2,3 +2,4 @@
 
 __attribute__((visibility("internal"))) void _init_multiplayer_raknet();
 __attribute__((visibility("internal"))) void _init_multiplayer_loading();
+__attribute__((visibility("internal"))) void _init_multiplayer_syncing();